Refactor Fake+Mock Objects, Use assertThat where possible
