+++ app/src/test/java/de/lukaspanni/opensourcestats/ContributionCountUnitTest.java
@@ -4,6 +4,8 @@ import org.junit.Test;
 
 import de.lukaspanni.opensourcestats.data.ContributionCount;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.not;
 import static org.junit.Assert.*;
 
 
@@ -12,27 +14,27 @@ public class ContributionCountUnitTest {
     @Test
     public void constructor_commits_smallerZero_correct() {
         ContributionCount testContributionCount = new ContributionCount(-2, 0, 0, 0);
-        assertNotEquals(-2, testContributionCount.getCommitCount());
-        assertEquals(0, testContributionCount.getCommitCount());
+        assertThat(testContributionCount.getCommitCount(), is(not(-2)));
+        assertThat(testContributionCount.getCommitCount(), is(0));
     }
 
     @Test
     public void constructor_issues_smallerZero_correct() {
         ContributionCount testContributionCount = new ContributionCount(0, -2, 0, 0);
-        assertNotEquals(-2, testContributionCount.getIssueCount());
-        assertEquals(0, testContributionCount.getIssueCount());
+        assertThat(testContributionCount.getIssueCount(), is(not(-2)));
+        assertThat(testContributionCount.getIssueCount(), is(0));
     }
 
     @Test
     public void constructor_pullRequests_smallerZero_correct() {
         ContributionCount testContributionCount = new ContributionCount(0, 0, -2, 0);
-        assertNotEquals(-2, testContributionCount.getPullRequestCount());
-        assertEquals(0, testContributionCount.getPullRequestCount());
+        assertThat(testContributionCount.getPullRequestCount(), is(not(-2)));
+        assertThat(testContributionCount.getPullRequestCount(), is(0));
     }
     @Test
     public void constructor_pullRequestReviews_smallerZero_correct() {
         ContributionCount testContributionCount = new ContributionCount(0, 0, 0,-2);
-        assertNotEquals(-2, testContributionCount.getPullRequestReviewCount());
-        assertEquals(0, testContributionCount.getPullRequestReviewCount());
+        assertThat(testContributionCount.getPullRequestReviewCount(), is(not(-2)));
+        assertThat(testContributionCount.getPullRequestReviewCount(), is(0));
     }
 }
