+++ app/src/test/java/de/lukaspanni/opensourcestats/TimeSpanUnitTest.java
@@ -6,6 +6,7 @@ import java.util.Date;
 
 import de.lukaspanni.opensourcestats.util.TimeSpan;
 
+import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
 
 
@@ -22,19 +23,19 @@ public class TimeSpanUnitTest {
         TimeSpan testTimeSpan = new TimeSpan(start, end);
         Date testTimeSpanStart = testTimeSpan.getStart();
         Date testTimeSpanEnd = testTimeSpan.getEnd();
-        assertEquals(testYear, testTimeSpanStart.getYear());
-        assertEquals(testMonth, testTimeSpanStart.getMonth());
-        assertEquals(testDateStart, testTimeSpanStart.getDate());
-        assertEquals(0, testTimeSpanStart.getHours());
-        assertEquals(0, testTimeSpanStart.getMinutes());
-        assertEquals(0, testTimeSpanStart.getSeconds());
+        assertThat(testTimeSpanStart.getYear(), is(testYear));
+        assertThat(testTimeSpanStart.getMonth(), is(testMonth));
+        assertThat(testTimeSpanStart.getDate(), is(testDateStart));
+        assertThat(testTimeSpanStart.getHours(), is(0));
+        assertThat(testTimeSpanStart.getMinutes(), is(0));
+        assertThat(testTimeSpanStart.getSeconds(), is(0));
         //Milliseconds are ignored!
-        assertEquals(testYear, testTimeSpanEnd.getYear());
-        assertEquals(testMonth, testTimeSpanEnd.getMonth());
-        assertEquals(testDateEnd, testTimeSpanEnd.getDate());
-        assertEquals(0, testTimeSpanEnd.getHours());
-        assertEquals(0, testTimeSpanEnd.getMinutes());
-        assertEquals(0, testTimeSpanEnd.getSeconds());
+        assertThat(testTimeSpanEnd.getYear(), is(testYear));
+        assertThat(testTimeSpanEnd.getMonth(), is(testMonth));
+        assertThat(testTimeSpanEnd.getDate(), is(testDateEnd));
+        assertThat(testTimeSpanEnd.getHours(), is(0));
+        assertThat(testTimeSpanEnd.getMinutes(), is(0));
+        assertThat(testTimeSpanEnd.getSeconds(), is(0));
         //again: Milliseconds ignored!
     }
